Refactor ClipboardCutEventProducer
